seeders

+++ database/migrations/2022_10_11_205143_create_emails_history.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,20 +12,20 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('emails_history', function (Blueprint $table) {
-
-            $table->uuid('id');
-            $table->string('code');
-            $table->string('name');
-            $table->string('email');
-            $table->string('phone');
-            $table->longText('message')->default('');
-            $table->string('type');
-            $table->string('status')->default('pending');
-            $table->timestamps();
-
-            $table->primary('id');
-        });
+        if (!Schema::hasTable('emails_history')) {
+            Schema::create('emails_history', function (Blueprint $table) {
+                $table->uuid('id');
+                $table->string('code');
+                $table->string('name');
+                $table->string('email');
+                $table->string('phone');
+                $table->longText('message')->default('');
+                $table->string('type');
+                $table->string('status')->default('pending');
+                $table->timestamps();
+                $table->primary('id');
+            });
+        }
     }
 
     /**
@@ -36,6 +35,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('emails_history');
+        if (Schema::hasTable('emails_history')) {
+            Schema::dropIfExists('emails_history');
+        }
     }
 };
