+++ database/migrations/2022_10_22_222714_create_email_notification.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,14 +12,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('emails_notifications', function (Blueprint $table) {
-            $table->uuid('id');
-            $table->string('name')->nullable(false);
-            $table->string('email')->nullable(false)->unique();
-            $table->timestamps();
+        if (!Schema::hasTable('emails_notifications')) {
+            Schema::create('emails_notifications', function (Blueprint $table) {
+                $table->uuid('id');
+                $table->string('name')->nullable(false);
+                $table->string('email')->nullable(false)->unique();
+                $table->timestamps();
 
-            $table->primary('id');
-        });
+                $table->primary('id');
+            });
+        }
     }
 
     /**
@@ -30,6 +31,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('emails_notifications');
+        if (Schema::hasTable('emails_notifications')) {
+            Schema::dropIfExists('emails_notifications');
+        }
     }
 };
